// src/widgets/LayoutHeader/model/navigationType.ts
// class LinkType {
// 	id: number;
// 	name: string;
// 	path: string;

// 	constructor(navId: number, navName: string, navPath: string) {
// 		this.id = navId;
// 		this.name = navName;
// 		this.path = navPath;
// 	}
// }

// export default LinkType;

export type linkType = {
	id: number;
	name: string;
	path: string;
};